Remove TradePeerTxInputValidator and TradePeerTxInputValidatorTest and move methods to DepositTxValidation
