+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/buyer_as_maker/BuyerAsMakerCreatesAndSignsDepositTx.java
@@ -24,8 +24,8 @@ import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.offer.Offer;
 import bisq.core.trade.model.bisq_v1.Trade;
 import bisq.core.trade.protocol.bisq_v1.model.TradingPeer;
-import bisq.core.trade.validation.TradePeerTxInputValidator;
 import bisq.core.trade.protocol.bisq_v1.tasks.TradeTask;
+import bisq.core.trade.validation.DepositTxValidation;
 
 import bisq.common.taskrunner.TaskRunner;
 
@@ -74,7 +74,7 @@ public class BuyerAsMakerCreatesAndSignsDepositTx extends TradeTask {
             Coin expectedTakersInputAmount = offer.getSellerSecurityDeposit()
                     .add(tradeAmount)
                     .add(trade.getTradeTxFee().multiply(2));
-            TradePeerTxInputValidator.validatePeersInputs(takerRawTransactionInputs,
+            DepositTxValidation.validatePeersInputs(takerRawTransactionInputs,
                     expectedTakersInputAmount,
                     walletService,
                     "Taker");
